refactor: Rename exercise type files and update imports
Renamed `weightTraining.ts` to `exerciseTypes.ts` and updated imports to reflect the change.  This improves the organization and clarity of the type definitions.  No breaking changes.
