+++ frontend/src/types/exercise/exerciseTypes.ts
